Actualizar tipos de ID en servicios y controladores: cambiar de Integer a Long en CineService, CineServiceImpl, y controladores relacionados.

+++ src/main/java/com/example/CineSpringBoot/services/CineService.java
@@ -2,5 +2,5 @@ package com.example.CineSpringBoot.services;
 
 import com.example.CineSpringBoot.entities.Cine;
 
-public interface CineService extends BaseService<Cine, Integer> {
+public interface CineService extends BaseService<Cine, Long> {
 }
